# backend/openapi/UKPS.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "UKPS.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/organisations": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "summary": "Retrieves all organisations available to the current user.",
        "operationId": "GetOrganisations",
        "responses": {
          "200": {
            "description": "The organisations were returned successfully.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationListDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationListDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationListDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organisation"
        ],
        "summary": "Creates a new organisation.",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganisationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganisationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganisationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{id}": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "summary": "Retrieves an organisation by its unique identifier.",
        "operationId": "GetOrganisationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier of the organisation to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The organisation was found and its details were returned.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "No organisation exists with the specified identifier.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organisation"
        ],
        "summary": "Updates the details of an existing organisation.",
        "operationId": "UpdateOrganisationDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier of the organisation to update.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "A token that can be used to cancel the operation.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganisationDetailsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganisationDetailsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganisationDetailsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The organisation was successfully updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "The supplied organisation details failed validation.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No organisation exists with the specified identifier.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{organisationId}/memberships/{membershipId}/deactivate": {
      "patch": {
        "tags": [
          "Organisation"
        ],
        "summary": "Deactivates an organisation membership.",
        "operationId": "DeactivateMembership",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "description": "The identifier of the organisation that owns the membership.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "membershipId",
            "in": "path",
            "description": "The identifier of the membership to deactivate.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The membership was successfully deactivated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not authorised to deactivate the membership, or the membership is their own.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The specified organisation membership could not be found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{organisationId}/memberships/{membershipId}/reactivate": {
      "patch": {
        "tags": [
          "Organisation"
        ],
        "summary": "Reactivates an organisation membership.",
        "operationId": "ReactivateMembership",
        "parameters": [
          {
            "name": "organisationId",
            "in": "path",
            "description": "The identifier of the organisation that owns the membership.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "membershipId",
            "in": "path",
            "description": "The identifier of the membership to reactivate.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The membership was successfully reactivated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationMembershipDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The specified organisation membership could not be found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganisationDto": {
        "required": [
          "organisationName",
          "headOfficeAddress",
          "headOfficeEmail",
          "headOfficeTelephone"
        ],
        "type": "object",
        "properties": {
          "organisationName": {
            "type": "string",
            "description": "Gets or sets the name of the organisation."
          },
          "countryOrRegion": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets or sets the country or region where the organisation is based, if applicable."
          },
          "headOfficeAddress": {
            "type": "string",
            "description": "Gets or sets the head office address of the organisation."
          },
          "headOfficeEmail": {
            "type": "string",
            "description": "Gets or sets the head office email address of the organisation."
          },
          "headOfficeTelephone": {
            "type": "string",
            "description": "Gets or sets the head office telephone number of the organisation."
          }
        },
        "description": "Represents the information required to create a new organisation."
      },
...
      "OrganisationListDto": {
        "required": [
          "organisationName",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Gets the unique identifier of the organisation.",
            "format": "int32"
          },
          "organisationName": {
            "type": "string",
            "description": "Gets the name of the organisation."
          }
        },
        "description": "Represents the ID and name of an organisation."
      },
      "OrganisationMembershipDto": {
        "required": [
          "id",
          "userId",
          "organisationId",
          "userRole",
          "status",
          "allowedPharmaceuticalEntity",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Gets the unique identifier of the organisation membership.",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "description": "Gets the unique identifier of the user associated with the membership.",
            "format": "int32"
          },
          "organisationId": {
            "type": "integer",
            "description": "Gets the unique identifier of the organisation associated with the membership.",
            "format": "int32"
          },
          "userRole": {
            "description": "Gets the role of the user within the organisation.",
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "description": "Gets the status of the user within the organisation.",
            "$ref": "#/components/schemas/UserOrgStatus"
          },
          "allowedPharmaceuticalEntity": {
            "description": "Gets the pharmaceutical entity that the user is allowed to access.",
            "$ref": "#/components/schemas/PharmaceuticalEntity"
          },
          "createdAt": {
            "type": "string",
            "description": "Gets the date and time when the organisation membership was created.",
            "format": "date-time"
          }
        },
        "description": "Represents the data transfer object for an organisation membership."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}